% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing.R
\name{BPM6_import_structure}
\alias{BPM6_import_structure}
\title{Import structural metadata for the BPM6 table}
\usage{
BPM6_import_structure(con, schema = "platform", keep_vintage = FALSE)
}
\arguments{
\item{con}{connection to database}

\item{schema}{schema name, defaults to "platform"}

\item{keep_vintage}{logical indicating whether to keep vintages, defaults to F}
}
\value{
nothing
}
\description{
Umbrella function that prepares and import all the metadata tables into
the database. It uses the functions from the UMARimportR package to
insert the data into the database.
}
