% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bpm6_levels}
\alias{bpm6_levels}
\title{BPM6 levels}
\format{
a dataframe with two columns:
\describe{
\item{code}{numeric code}
\item{description}{level description}
}
}
\usage{
bpm6_levels
}
\description{
dataframe with code and description for the 1531 levels in the 6th edition
Balance of payments and international investment position manyal.
}
\keyword{datasets}
