% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_excel.R
\name{format_period_from_date}
\alias{format_period_from_date}
\title{Format Excel date serial to period ID}
\usage{
format_period_from_date(excel_date)
}
\arguments{
\item{excel_date}{Numeric Excel date serial}
}
\value{
Character period ID in format "YYYYMM"
}
\description{
Format Excel date serial to period ID
}
\keyword{internal}
